rearrange server files
