+++ services/server/cg_server/db_seed/__init__.py
@@ -1,0 +1,9 @@
+# coding: utf-8
+
+"""Database seeding
+
+Author: Albert Chen - Vector Engineering Team (chena@broadinstitute.org)
+"""
+
+from .seed import seed_database
+from .insert_sequences import insert_sequences
