+++ services/server/cg_server/download/__init__.py
@@ -1,0 +1,10 @@
+# coding: utf-8
+
+"""Download data functions
+
+Author: Albert Chen - Vector Engineering Team (chena@broadinstitute.org)
+"""
+
+from .genomes import download_genomes
+from .metadata import download_metadata
+from .snvs import download_snvs
